resume agent/resume_agent/requirements.txt
openai
flask
python-dotenv
pdfplumber
fpdf
flask-cors